fix(dataset): keep the band type clip computed when a bound is a numpy scalar
clip decided the result's type with _mask_dtype — the same judgement where and fillna
make — cast the values to it, and then let np.clip promote the answer back under NEP 50.
A bound is a numpy scalar whenever it came from a percentile, a mean or a quantile, which
is the commonest way clip is called:

    # before
    >>> floats.dtype                                     # float32 band
    ['float32']
    >>> floats.clip(2.0, 6.0).dtype
    ['float32']
    >>> floats.clip(np.float64(2.0), np.float64(6.0)).dtype
    ['float64']
    >>> low, high = np.percentile([1.0, 5.0, 9.0], [10, 90]); floats.clip(low, high).dtype
    ['float64']
    >>> bytes_.clip(np.int64(2), np.int64(6)).dtype      # uint8 band -> GDAL Int64
    ['int64']

    # after
    >>> floats.clip(np.float64(2.0), np.float64(6.0)).dtype
    ['float32']
    >>> floats.clip(low, high).dtype
    ['float32']
    >>> bytes_.clip(np.int64(2), np.int64(6)).dtype
    ['uint8']

The values are untouched ([2.0, 5.0, 6.0]) and the widening judgement still fires when the
bound genuinely does not fit: a uint8 band clipped at 300 still comes back float64 holding
[300.0, 300.0, 300.0].

round and astype do not share the defect — round's np.round preserves the input dtype and
astype allocates in the target type — and isin builds uint8 flags. where already cast its
np.where output for this reason, which is where the pattern comes from.

+++ src/pyramids/dataset/engines/analysis.py
@@ -2805,7 +2805,13 @@ class Analysis(_Engine["Dataset"]):
         for bound in (lower, upper):
             if bound is not None:
                 dtype = np.promote_types(dtype, _mask_dtype(values.dtype, bound))
-        clipped = np.clip(values.astype(dtype, copy=False), lower, upper)
+        # Cast back after clipping, not only before it: a bound that arrives as a numpy
+        # scalar — every percentile, mean or quantile does — promotes the result under NEP
+        # 50, so a float32 band came back float64 and a uint8 band int64. `where` casts its
+        # own output for the same reason.
+        clipped = np.clip(values.astype(dtype, copy=False), lower, upper).astype(
+            dtype, copy=False
+        )
         return self._identified(
             self._rebuilt(_regapped(clipped, domain, declared), declared)
         )
